swap request code started

+++ backend/DutyMatrix/DutyMatrix/src/main/java/com/DutyMatrix/dto/UserRegisterDTO.java
@@ -18,5 +18,5 @@ public class UserRegisterDTO {
 	private UserRank urank;
 	private UserRole urole;
 
-	private int station_id;
+	private Long station_id;
 }
