+++ backend/DutyMatrix/DutyMatrix/src/main/java/com/DutyMatrix/repositories/StationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.DutyMatrix.pojo.Station;
 
-public interface StationRepository extends JpaRepository<Station, Integer> {
+public interface StationRepository extends JpaRepository<Station, Long> {
 
 }
